Fix warning about python version.
Currently we can see following warning in CI/CD:
```
warning: The `requires-python` specifier (`~=3.10`) in `kaggle-bitcoin` uses the tilde specifier (`~=`) without a patch version. This will be interpreted as `>=3.10, <4`. Did you mean `~=3.10.0` to constrain the version as `>=3.10.0, <3.11`? We recommend only using the tilde specifier with a patch version to avoid ambiguity.
```

Lets fix it.